Move hooks to lib
