+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog/ElementDialog.tsx
@@ -6,10 +6,10 @@ import {
   Dialog,
   ListBox,
 } from "app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
-import { useElementOptions } from "hooks";
+import { useElementOptions } from "@lib/hooks";
 import { ElementOption, ElementOptionsFilter } from "@clientComponents/form-builder/types";
 import { Button } from "@clientComponents/globals";
-import { Groups } from "hooks/useElementOptions";
+import { Groups } from "@lib/hooks/useElementOptions";
 import { ElementFilters } from "./ElementFilters";
 
 export type SelectedGroupState = {
