+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/layout.tsx
@@ -7,8 +7,7 @@ import { AccessControlError, createAbility } from "@lib/privileges";
 import { getFullTemplateByID } from "@lib/templates";
 import { redirect } from "next/navigation";
 import { TemplateStoreProvider } from "@clientComponents/form-builder/store";
-import { TemplateApiProvider } from "hooks";
-import { RefStoreProvider } from "@lib/hooks/useRefStore";
+import { TemplateApiProvider, RefStoreProvider } from "@lib/hooks";
 
 export default async function Layout({
   children,
