+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/preview/Preview.tsx
@@ -17,7 +17,7 @@ import { BackArrowIcon } from "@serverComponents/icons";
 import Brand from "@clientComponents/globals/Brand";
 import { useIsFormClosed } from "@lib/hooks/useIsFormClosed";
 import { GCFormsProvider } from "@lib/hooks/useGCFormContext";
-import { useRehydrate } from "hooks";
+import { useRehydrate } from "@lib/hooks";
 import Skeleton from "react-loading-skeleton";
 
 export const Preview = () => {
