--- hooks/index.ts
@@ -1,11 +1,0 @@
-export { useAllowPublish } from "./useAllowPublish";
-export { useElementOptions } from "./useElementOptions";
-export { useTemplateApi } from "./useTemplateApi";
-export { useActivePathname } from "./useActivePathname";
-export { useTemplateStatus } from "./useTemplateStatus";
-export { TemplateApiProvider, useTemplateContext } from "./useTemplateContext";
-export { useIsWithin } from "./useIsWithin";
-export { usePanelActions } from "./usePanelActions";
-export { useHandleAdd } from "./useHandleAdd";
-export { useIsAdminUser } from "./useIsAdminUser";
-export { useRehydrate } from "../components/clientComponents/form-builder/store/useTemplateStore";
