+++ lib/hooks/index.ts
@@ -5,3 +5,15 @@ export { useFormTimer } from "./useFormTimer";
 export { useAccessControl, AccessControlProvider } from "./useAccessControl";
 export { useIsFormClosed } from "./useIsFormClosed";
 export { useFormValuesChanged } from "./useValueChanged";
+export { useAllowPublish } from "./useAllowPublish";
+export { useElementOptions } from "./useElementOptions";
+export { useTemplateApi } from "./useTemplateApi";
+export { useActivePathname } from "./useActivePathname";
+export { useTemplateStatus } from "./useTemplateStatus";
+export { TemplateApiProvider, useTemplateContext } from "./useTemplateContext";
+export { useIsWithin } from "./useIsWithin";
+export { usePanelActions } from "./usePanelActions";
+export { useHandleAdd } from "./useHandleAdd";
+export { useIsAdminUser } from "./useIsAdminUser";
+export { useRehydrate } from "../../components/clientComponents/form-builder/store/useTemplateStore";
+export { RefStoreProvider } from "./useRefStore";
