+++ lib/hooks/useElementOptions.tsx
@@ -35,14 +35,14 @@ import {
   FileInput,
   Departments,
   Combobox,
-} from "../app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog";
+} from "../../app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog";
 
 import { useIsAdminUser } from "./useIsAdminUser";
 
 import {
   ElementOptionsFilter,
   ElementOption,
-} from "../components/clientComponents/form-builder/types";
+} from "../../components/clientComponents/form-builder/types";
 import { useFlag } from "@lib/hooks";
 
 export enum Groups {
