+++ lib/hooks/usePanelActions.tsx
