+++ lib/hooks/useTemplateContext.tsx
@@ -3,14 +3,14 @@ import React, { createContext, useState, useContext, useRef, useCallback } from 
 import {
   useTemplateStore,
   useSubscibeToTemplateStore,
-} from "../components/clientComponents/form-builder/store";
+} from "../../components/clientComponents/form-builder/store";
 import { useTemplateApi } from ".";
 import { useTranslation } from "@i18n/client";
 import { logMessage } from "@lib/logger";
 import { useSession } from "next-auth/react";
-import { toast } from "../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/Toast";
+import { toast } from "../../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/Toast";
 import { StyledLink } from "@clientComponents/globals";
-import { DownloadFileButton } from "../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
+import { DownloadFileButton } from "../../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
 
 interface TemplateApiType {
   error: string | null | undefined;
